code/requirements.txt
Django>=5.1,<5.2
django-silk==5.1.0
passlib==1.7.4
psycopg2-binary>=2.9,<3.0
django-ninja==1.3.0
Pillow>=10.0,<11.0
pydantic[email]
PyJWT==2.8.0



